refactor: correct nav style
