+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom'
 import { Header } from './header/Header'
-import { Footer } from './header/Footer'
+import { Footer } from './footer/Footer'
 
 const Layout = () => {
   return (
